# api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ApiResponseBatchGenerateResult" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/BatchGenerateResult"
          },
          "error" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseBillListResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/BillListResponse"
          },
          "error" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
      "ApiResponseBillResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/BillResponse"
          },
          "error" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseChargeConfigResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/ChargeConfigResponse"
          },
          "error" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseLeaseResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/LeaseResponse"
          },
          "error" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ApiResponsePageChargeConfigResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/PageChargeConfigResponse"
          },
          "error" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ApiResponsePageLeaseResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/PageLeaseResponse"
          },
          "error" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseVoid" : {
        "properties" : {
          "data" : { },
          "error" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "BatchGenerateBillRequest" : {
        "properties" : {
          "billingMonth" : {
            "pattern" : "\\d{4}-\\d{2}",
            "type" : "string"
          },
          "blockId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "dueDate" : {
            "format" : "date",
            "type" : "string"
          },
          "propertyId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "required" : [ "billingMonth", "dueDate", "propertyId" ],
        "type" : "object"
      },
      "BatchGenerateFailure" : {
        "properties" : {
          "leaseId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "unitNumber" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "BatchGenerateResult" : {
        "properties" : {
          "failed" : {
            "items" : {
              "$ref" : "#/components/schemas/BatchGenerateFailure"
            },
            "type" : "array"
          },
          "succeeded" : {
            "items" : {
              "$ref" : "#/components/schemas/BillResponse"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "BillListResponse" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/BillResponse"
            },
            "type" : "array"
          },
          "metrics" : {
            "$ref" : "#/components/schemas/BillMetricsDTO"
          },
          "number" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "BillMetricsDTO" : {
        "properties" : {
          "pendingDraftsCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "publishedCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalExpectedRevenue" : {
            "type" : "number"
          }
        },
        "type" : "object"
      },
...
      "BillResponse" : {
        "properties" : {
          "billingMonth" : {
            "type" : "string"
          },
          "blockId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "blockName" : {
            "type" : "string"
          },
          "charges" : {
            "items" : {
              "$ref" : "#/components/schemas/ChargeResponse"
            },
            "type" : "array"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "dueDate" : {
            "format" : "date",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "paidAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "enum" : [ "PENDING", "PUBLISHED", "PAID", "OVERDUE", "PARTIALLY_PAID" ],
            "type" : "string"
          },
          "tenantName" : {
            "type" : "string"
          },
          "totalAmount" : {
            "type" : "number"
          },
          "unitNumber" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ChargeConfigRequest" : {
        "properties" : {
          "applySalesTax" : {
            "type" : "boolean"
          },
          "autoCarryForward" : {
            "type" : "boolean"
          },
          "baseRate" : {
            "type" : "number"
          },
          "billingFrequency" : {
            "enum" : [ "MONTHLY", "ANNUAL", "WEEKLY", "ONE_TIME" ],
            "type" : "string"
          },
          "calculationStrategy" : {
            "enum" : [ "FIXED_RATE", "METERED" ],
            "type" : "string"
          },
          "chargeName" : {
            "type" : "string"
          },
          "lateFeePercentage" : {
            "type" : "number"
          },
          "propertyId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "unitType" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ChargeConfigResponse" : {
        "properties" : {
          "applySalesTax" : {
            "type" : "boolean"
          },
          "autoCarryForward" : {
            "type" : "boolean"
          },
          "baseRate" : {
            "type" : "number"
          },
          "billingFrequency" : {
            "enum" : [ "MONTHLY", "ANNUAL", "WEEKLY", "ONE_TIME" ],
            "type" : "string"
          },
          "calculationStrategy" : {
            "enum" : [ "FIXED_RATE", "METERED" ],
            "type" : "string"
          },
          "chargeName" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "isActive" : {
            "type" : "boolean"
          },
          "lateFeePercentage" : {
            "type" : "number"
          },
          "propertyId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "unitType" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ChargeResponse" : {
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "CreateLeaseRequest" : {
        "properties" : {
          "bookingId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "monthlyRentAmount" : {
            "type" : "number"
          },
          "moveInDate" : {
            "format" : "date",
            "type" : "string"
          },
          "moveOutDate" : {
            "format" : "date",
            "type" : "string"
          },
          "securityDeposit" : {
            "type" : "number"
          },
          "splitStrategy" : {
            "enum" : [ "FULL_UNIT", "PER_OCCUPANT", "CUSTOM" ],
            "type" : "string"
          },
          "status" : {
            "enum" : [ "Active", "Ended" ],
            "type" : "string"
          },
          "unitId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "userId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "required" : [ "monthlyRentAmount", "moveInDate", "securityDeposit", "splitStrategy", "unitId" ],
        "type" : "object"
      },
...
      "GenerateBillRequest" : {
        "properties" : {
          "billingMonth" : {
            "pattern" : "\\d{4}-\\d{2}",
            "type" : "string"
          },
          "dueDate" : {
            "format" : "date",
            "type" : "string"
          },
          "leaseId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "required" : [ "billingMonth", "dueDate", "leaseId" ],
        "type" : "object"
      },
...
      "PageChargeConfigResponse" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/ChargeConfigResponse"
            },
            "type" : "array"
          },
          "empty" : {
            "type" : "boolean"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "number" : {
            "format" : "int32",
            "type" : "integer"
          },
          "numberOfElements" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "PageLeaseResponse" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/LeaseResponse"
            },
            "type" : "array"
          },
          "empty" : {
            "type" : "boolean"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "number" : {
            "format" : "int32",
            "type" : "integer"
          },
          "numberOfElements" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "Pageable" : {
        "properties" : {
          "page" : {
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "minimum" : 1,
            "type" : "integer"
          },
          "sort" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "SortObject" : {
        "properties" : {
          "empty" : {
            "type" : "boolean"
          },
          "sorted" : {
            "type" : "boolean"
          },
          "unsorted" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "JWT access token from `POST /api/v1/auth/login` or `POST /api/v1/auth/signup`",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "contact" : {
      "email" : "dev@tenantliving.com",
      "name" : "Tenant Living Team"
    },
    "description" : "## Overview\nBackend API for the Tenant Living property management system.\n\n## Authentication\n- Obtain **`accessToken`** and **`refreshToken`** from **`POST /api/v1/auth/login`** or **`POST /api/v1/auth/signup`**.\n- Call protected endpoints with header **`Authorization`** value **`Bearer`** followed by the access token.\n- Rotate access with **`POST /api/v1/auth/refresh`** (send refresh token in JSON body).\n- Introspect a JWT without calling a protected route: **`POST /api/v1/auth/validate`**.\n\n## Errors\nValidation and business errors use **`application/json`** body **`ApiError`** (see schema).\n\n## Roles\n| Role | Description |\n|------|-------------|\n| SUPER_ADMIN | Primary landlord |\n| ADMIN | Co-owner / manager |\n| PROPERTY_STAFF | Operations staff |\n| USER | Tenant (default for self-signup) |\n",
    "license" : {
      "name" : "Proprietary",
      "url" : "https://tenantliving.com"
    },
    "title" : "Tenant Living API",
    "version" : "1.0.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/finance/bills" : {
      "get" : {
        "operationId" : "list_1",
        "parameters" : [ {
          "in" : "query",
          "name" : "propertyId",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "billingMonth",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "enum" : [ "PENDING", "PUBLISHED", "PAID", "OVERDUE", "PARTIALLY_PAID" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "search",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "pageable",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseBillListResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "bill-controller" ]
      }
    },
    "/api/v1/finance/bills/batch-generate" : {
      "post" : {
        "operationId" : "batchGenerate",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BatchGenerateBillRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseBatchGenerateResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "rent-generation-controller" ]
      }
    },
...
    "/api/v1/finance/bills/generate" : {
      "post" : {
        "operationId" : "generate",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GenerateBillRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseBillResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "rent-generation-controller" ]
      }
    },
...
    "/api/v1/finance/charge-configs" : {
      "post" : {
        "operationId" : "createChargeConfig",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChargeConfigRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseChargeConfigResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "charge-config-controller" ]
      }
    },
    "/api/v1/finance/charge-configs/property/{propertyId}" : {
      "get" : {
        "operationId" : "getChargesForProperty",
        "parameters" : [ {
          "in" : "path",
          "name" : "propertyId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "includeInactive",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          }
        }, {
          "in" : "query",
          "name" : "pageable",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePageChargeConfigResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "charge-config-controller" ]
      }
    },
    "/api/v1/finance/charge-configs/{id}" : {
      "delete" : {
        "operationId" : "deactivateChargeConfig",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "charge-config-controller" ]
      },
      "get" : {
        "operationId" : "getChargeConfigById",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseChargeConfigResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "charge-config-controller" ]
      },
      "put" : {
        "operationId" : "updateChargeConfig",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChargeConfigRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseChargeConfigResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "charge-config-controller" ]
      }
    },
    "/api/v1/finance/charge-configs/{id}/permanent" : {
      "delete" : {
        "operationId" : "deleteChargeConfigPermanently",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "charge-config-controller" ]
      }
    },
    "/api/v1/finance/charge-configs/{id}/reactivate" : {
      "post" : {
        "operationId" : "reactivateChargeConfig",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "charge-config-controller" ]
      }
    },
    "/api/v1/finance/leases" : {
      "get" : {
        "operationId" : "getActiveLeasesByProperty",
        "parameters" : [ {
          "in" : "query",
          "name" : "propertyId",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "blockId",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "pageable",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePageLeaseResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "lease-controller" ]
      },
      "post" : {
        "operationId" : "create_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateLeaseRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseLeaseResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "lease-controller" ]
      }
    },
...
    "/api/v1/finance/leases/{leaseId}/bills" : {
      "get" : {
        "operationId" : "listForLease",
        "parameters" : [ {
          "in" : "path",
          "name" : "leaseId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "billingMonth",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "enum" : [ "PENDING", "PUBLISHED", "PAID", "OVERDUE", "PARTIALLY_PAID" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "pageable",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseBillListResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "lease-bill-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Local (default dev port)",
    "url" : "http://localhost:8080"
  } ]
}