# backend/src/main/java/com/livic/core/finance/controller/ChargeConfigController.java
package com.livic.core.finance.controller;

import com.livic.platform.common.response.ApiResponse;
import com.livic.core.finance.dto.ChargeConfigRequest;
import com.livic.core.finance.dto.ChargeConfigResponse;
import com.livic.core.finance.service.interfaces.ChargeConfigQueryService;
import com.livic.core.finance.service.interfaces.ChargeConfigService;
import lombok.RequiredArgsConstructor;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.data.web.PageableDefault;
import org.springframework.http.ResponseEntity;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.web.bind.annotation.*;

import java.util.UUID;

@RestController
@RequestMapping("/api/v1/finance/charge-configs")
@RequiredArgsConstructor
public class ChargeConfigController {

    private final ChargeConfigService chargeConfigService;
    private final ChargeConfigQueryService chargeConfigQueryService;

    @PostMapping
    @PreAuthorize("@authorizationService.hasPermission(#request.propertyId, 'CHARGE_CONFIG_MANAGE')")
    public ResponseEntity<ApiResponse<ChargeConfigResponse>> createChargeConfig(@RequestBody ChargeConfigRequest request) {
        ChargeConfigResponse response = chargeConfigService.createChargeConfig(request);
        return ResponseEntity.ok(ApiResponse.success(response));
    }

    @PutMapping("/{id}")
    @PreAuthorize("@authorizationService.hasPermission(T(com.livic.core.finance.security.FinanceResources).CHARGE_CONFIG, #id, 'CHARGE_CONFIG_MANAGE')")
    public ResponseEntity<ApiResponse<ChargeConfigResponse>> updateChargeConfig(@PathVariable UUID id, @RequestBody ChargeConfigRequest request) {
        ChargeConfigResponse response = chargeConfigService.updateChargeConfig(id, request);
        return ResponseEntity.ok(ApiResponse.success(response));
    }

    @DeleteMapping("/{id}")
    @PreAuthorize("@authorizationService.hasPermission(T(com.livic.core.finance.security.FinanceResources).CHARGE_CONFIG, #id, 'CHARGE_CONFIG_MANAGE')")
    public ResponseEntity<ApiResponse<Void>> deactivateChargeConfig(@PathVariable UUID id) {
        chargeConfigService.deactivateChargeConfig(id);
        return ResponseEntity.ok(ApiResponse.success(null));
    }

    @PostMapping("/{id}/reactivate")
    @PreAuthorize("@authorizationService.hasPermission(T(com.livic.core.finance.security.FinanceResources).CHARGE_CONFIG, #id, 'CHARGE_CONFIG_MANAGE')")
    public ResponseEntity<ApiResponse<Void>> reactivateChargeConfig(@PathVariable UUID id) {
        chargeConfigService.reactivateChargeConfig(id);
        return ResponseEntity.ok(ApiResponse.success(null));
    }

    @DeleteMapping("/{id}/permanent")
    @PreAuthorize("@authorizationService.hasPermission(T(com.livic.core.finance.security.FinanceResources).CHARGE_CONFIG, #id, 'CHARGE_CONFIG_MANAGE')")
    public ResponseEntity<ApiResponse<Void>> deleteChargeConfigPermanently(@PathVariable UUID id) {
        chargeConfigService.deleteChargeConfigPermanently(id);
        return ResponseEntity.ok(ApiResponse.success(null));
    }

    @GetMapping("/property/{propertyId}")
    // Meter readings and billing worksheets pick a charge from this list, so their viewers can read it too.
    @PreAuthorize("@authorizationService.hasAnyPermission(#propertyId, 'CHARGE_CONFIG_VIEW', 'METER_READING_VIEW', 'BILLING_WORKSHEET_VIEW')")
    public ResponseEntity<ApiResponse<Page<ChargeConfigResponse>>> getChargesForProperty(
            @PathVariable UUID propertyId,
            @RequestParam(required = false, defaultValue = "false") boolean includeInactive,
            @PageableDefault(size = 20) Pageable pageable
    ) {
        Page<ChargeConfigResponse> responses = chargeConfigQueryService.getChargesForProperty(propertyId, includeInactive, pageable);
        return ResponseEntity.ok(ApiResponse.success(responses));
    }

    @GetMapping("/{id}")
    @PreAuthorize("@authorizationService.hasPermission(T(com.livic.core.finance.security.FinanceResources).CHARGE_CONFIG, #id, 'CHARGE_CONFIG_VIEW')")
    public ResponseEntity<ApiResponse<ChargeConfigResponse>> getChargeConfigById(@PathVariable UUID id) {
        ChargeConfigResponse response = chargeConfigQueryService.getChargeConfigById(id);
        return ResponseEntity.ok(ApiResponse.success(response));
    }
}